# backend/src/main/resources/openapi/tracex-api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:62436
  description: Generated server url
paths:
...
  /api/v1/import/{id}/rollback:
    post:
      tags:
      - import-controller
      operationId: rollbackImport
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseImportRollbackResponseDto'
  /api/v1/import/validate:
    post:
      tags:
      - import-controller
      operationId: validateImport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportValidateRequestDto'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseImportValidateResponseDto'
  /api/v1/import/map-headers:
    post:
      tags:
      - import-controller
      operationId: mapHeaders
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportMapHeadersRequestDto'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseImportMapHeadersResponseDto'
  /api/v1/import/commit:
    post:
      tags:
      - import-controller
      operationId: commitImport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportCommitRequestDto'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseImportCommitResponseDto'
  /api/v1/batches:
    get:
      tags:
      - batch-controller
      operationId: getAllBatches
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: sku
        in: query
        required: false
        schema:
          type: string
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: sort
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchPageResult'
    post:
      tags:
      - batch-controller
      operationId: createBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCreateDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchDetailDto'
...
  /api/v1/batches/{id}/restore:
    patch:
      tags:
      - batch-controller
      operationId: restoreBatch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchDetailDto'
  /api/v1/batches/{id}/raw-material:
    patch:
      tags:
      - batch-controller
      operationId: updateRawMaterial
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchRawMaterialDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchDetailDto'
  /api/v1/batches/{id}/note:
    patch:
      tags:
      - batch-controller
      operationId: updateNote
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchNoteDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchDetailDto'
  /api/v1/batches/{id}/dispatch:
    patch:
      tags:
      - batch-controller
      operationId: dispatchBatch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDispatchDto'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchDetailDto'
...
  /api/v1/inspections/batch/{batchId}:
    get:
      tags:
      - inspection-controller
      operationId: listByBatch
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListInspection'
...
  /api/v1/import/{id}:
    get:
      tags:
      - import-controller
      operationId: getImportJob
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseImportJobDetailDto'
...
  /api/v1/batches/{id}:
    get:
      tags:
      - batch-controller
      operationId: getBatchById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchDetailDto'
    delete:
      tags:
      - batch-controller
      operationId: archiveBatch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchDetailDto'
  /api/v1/batches/{id}/scans:
    get:
      tags:
      - batch-controller
      operationId: getBatchScans
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchScansDto'
  /api/v1/batches/{id}/qr:
    get:
      tags:
      - batch-controller
      operationId: getBatchQr
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBatchQrDto'
...
components:
  schemas:
...
    FieldErrorDto:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Inspection:
      type: object
      properties:
        id:
          type: string
        batchId:
          type: string
        batchCode:
          type: string
        productName:
          type: string
        sku:
          type: string
        status:
          type: string
        rating:
          type: integer
          format: int32
        checklist:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        findings:
          type: string
        recommendation:
          type: string
        inspectedBy:
          $ref: '#/components/schemas/InspectedBy'
        createdAt:
          type: string
          format: date-time
        isLatest:
          type: boolean
    ApiResponseImportRollbackResponseDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/ImportRollbackResponseDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    ImportRollbackResponseDto:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
        archived:
          type: integer
          format: int32
        alreadyArchived:
          type: integer
          format: int32
        rolledBackAt:
          type: string
          format: date-time
        rolledBackBy:
          type: string
    ImportValidateRequestDto:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        rowOffset:
          type: integer
          format: int32
        priorKeys:
          type: array
          items:
            type: string
    ApiResponseImportValidateResponseDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/ImportValidateResponseDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
...
    ImportPreviewRowDto:
      type: object
      properties:
        rowNumber:
          type: integer
          format: int32
        verdict:
          type: string
        reason:
          type: string
        product:
          $ref: '#/components/schemas/ImportPreviewProductDto'
        sourceLotCode:
          type: string
        village:
          type: string
        quantityProduced:
          type: integer
          format: int32
        unit:
          type: string
        yieldPercent:
          type: number
          format: double
        packDate:
          type: string
        insertKey:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ImportFieldErrorDto'
    ImportValidateResponseDto:
      type: object
      properties:
        summary:
          $ref: '#/components/schemas/ImportValidateSummaryDto'
        preview:
          type: array
          items:
            $ref: '#/components/schemas/ImportPreviewRowDto'
    ImportValidateSummaryDto:
      type: object
      properties:
        total:
          type: integer
          format: int32
        insert:
          type: integer
          format: int32
        skip:
          type: integer
          format: int32
        error:
          type: integer
          format: int32
    ImportMapHeadersRequestDto:
      type: object
      properties:
        headers:
          type: array
          items:
            type: string
    ApiResponseImportMapHeadersResponseDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/ImportMapHeadersResponseDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    ImportMapHeadersResponseDto:
      type: object
      properties:
        mapping:
          type: object
          additionalProperties:
            type: string
        unmappedRequired:
          type: array
          items:
            type: string
    ImportCommitRequestDto:
      type: object
      properties:
        jobId:
          type: string
        fileName:
          type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        rowOffset:
          type: integer
          format: int32
        totalRows:
          type: integer
          format: int32
        isFinal:
          type: boolean
    ApiResponseImportCommitResponseDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/ImportCommitResponseDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    ImportCommitResponseDto:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
        chunkInserted:
          type: integer
          format: int32
        chunkSkipped:
          type: integer
          format: int32
        chunkErrored:
          type: integer
          format: int32
        totals:
          $ref: '#/components/schemas/ImportCommitTotalsDto'
        batches:
          type: array
          items:
            $ref: '#/components/schemas/ImportedBatchSummaryDto'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/RowError'
        rowErrorsTruncated:
          type: boolean
    ImportCommitTotalsDto:
      type: object
      properties:
        processedRows:
          type: integer
          format: int32
        totalRows:
          type: integer
          format: int32
        inserted:
          type: integer
          format: int32
        skipped:
          type: integer
          format: int32
        errored:
          type: integer
          format: int32
...
    BatchCreateDto:
      required:
      - farmerName
      - packDate
      - productId
      - sourceLotCode
      - unit
      - village
      type: object
      properties:
        productId:
          type: string
        sourceLotCode:
          maxLength: 100
          minLength: 0
          type: string
        farmerName:
          maxLength: 200
          minLength: 0
          type: string
        village:
          maxLength: 200
          minLength: 0
          type: string
        quantityProduced:
          minimum: 1
          type: integer
          format: int32
        unit:
          pattern: Kg|Units|Liters
          type: string
        yieldPercent:
          maximum: 100.0
          exclusiveMaximum: false
          minimum: 0.0
          exclusiveMinimum: false
          type: number
          format: double
        packDate:
          type: string
          format: date
          example: 2026-10-09
        expiryDate:
          type: string
          format: date
          example: 2026-10-09
        traceabilityNote:
          maxLength: 2000
          minLength: 0
          type: string
    ApiResponseBatchDetailDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/BatchDetailDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    BatchDetailDto:
      type: object
      properties:
        id:
          type: string
        batchCode:
          type: string
        productName:
          type: string
        sku:
          type: string
        sourceLotCode:
          type: string
        farmerName:
          type: string
        village:
          type: string
        quantityProduced:
          type: integer
          format: int32
        unit:
          type: string
        yieldPercent:
          type: number
          format: double
        packDate:
          type: string
          format: date
          example: 2026-10-09
        expiryDate:
          type: string
          format: date
          example: 2026-10-09
        dataSource:
          type: string
        shelfLifeSource:
          type: string
        lifecycleState:
          type: string
        status:
          type: string
          description: Computed batch freshness or lifecycle status
          enum:
          - EXPIRED
          - URGENT
          - WARNING
          - READY
          - EXCEPTION
          - DISPATCHED
        daysUntilExpiry:
          type: integer
          format: int64
        priorityScore:
          type: number
          format: double
        qualityCheck:
          $ref: '#/components/schemas/QualityCheck'
        rank:
          type: integer
          format: int32
        exceptionReason:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        traceabilityNote:
          type: string
        noteHistory:
          type: array
          items:
            $ref: '#/components/schemas/NoteHistoryEntry'
        dispatchHistory:
          type: array
          items:
            $ref: '#/components/schemas/DispatchHistoryEntry'
        deletedAt:
          type: string
          format: date-time
        deletedBy:
          type: string
        deleteNote:
          type: string
        dispatchDate:
          type: string
          format: date
          example: 2026-10-09
        buyerName:
          type: string
        warning:
          type: string
        qrAbsoluteUrl:
          type: string
        deleted:
          type: boolean
...
    BatchRawMaterialDto:
      type: object
      properties:
        farmerName:
          maxLength: 200
          minLength: 0
          type: string
        village:
          maxLength: 200
          minLength: 0
          type: string
        sourceLotCode:
          maxLength: 100
          minLength: 0
          type: string
        quantityProduced:
          minimum: 1
          type: integer
          format: int32
        unit:
          pattern: Kg|Units|Liters
          type: string
        yieldPercent:
          maximum: 100.0
          exclusiveMaximum: false
          minimum: 0.0
          exclusiveMinimum: false
          type: number
          format: double
        expiryDate:
          type: string
          format: date
          example: 2026-10-09
    BatchNoteDto:
      required:
      - note
      type: object
      properties:
        note:
          maxLength: 2000
          minLength: 0
          type: string
    BatchDispatchDto:
      required:
      - buyerName
      type: object
      properties:
        buyerName:
          maxLength: 200
          minLength: 0
          type: string
        dispatchDate:
          type: string
          format: date
          example: 2026-10-09
        overrideReason:
          maxLength: 500
          minLength: 0
          type: string
...
    ApiResponseListInspection:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Inspection'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
...
    ApiResponseImportJobDetailDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/ImportJobDetailDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    ImportJobDetailDto:
      type: object
      properties:
        id:
          type: string
        fileName:
          type: string
        entity:
          type: string
        status:
          type: string
        totalRows:
          type: integer
          format: int32
        processedRows:
          type: integer
          format: int32
        inserted:
          type: integer
          format: int32
        updated:
          type: integer
          format: int32
        skipped:
          type: integer
          format: int32
        errored:
          type: integer
          format: int32
        rowErrorsTruncated:
          type: boolean
        createdBy:
          type: string
        createdByRole:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        rolledBackAt:
          type: string
          format: date-time
        rolledBackBy:
          type: string
        insertedBatchIds:
          type: array
          items:
            type: string
        rowErrors:
          type: array
          items:
            $ref: '#/components/schemas/RowError'
...
    ApiResponseBatchPageResult:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/BatchPageResult'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    BatchPageResult:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BatchSummaryDto'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        count:
          type: integer
          format: int32
    ApiResponseBatchScansDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/BatchScansDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    BatchScansDto:
      type: object
      properties:
        total:
          type: integer
          format: int64
        lastScannedAt:
          type: string
          format: date-time
        byDevice:
          type: object
          additionalProperties:
            type: integer
            format: int64
        bySource:
          type: object
          additionalProperties:
            type: integer
            format: int64
    ApiResponseBatchQrDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/BatchQrDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    BatchQrDto:
      type: object
      properties:
        qrCodeDataUrl:
          type: string
        qrAbsoluteUrl:
          type: string
...